requirements.txt
Django==1.11.5
django-autocomplete-light==3.2.10
django-common-helpers==0.9.1
django-datatable==0.3.1
django-jsonfield==1.0.1
django-model-utils==3.0.0
django-notify-x==0.1.7
jsonfield==2.0.2
pytz==2017.2
six==1.11.0
